use pushed_at over updated_at
It seems `updated_at` refers to the last change to the repo, annoyingly including adding of topics. I was wondering why some of my 'most recent updates' were to repos I hadn't really touched in a while (beyond adding topics). http://stackoverflow.com/a/15922637/4168169

There is a give and take though - `updated_at` updates on push to any branch, but `pushed_at` only responds to pushes to `master`.

Also minor: `gh` isn't on CRAN so I thought it worthwhile to mention the source, and I was having odd issues with digits from the `sprintf`. I was originally trying to re-sort the `Age` column with the `DT::datatable` button before I realised the units prevents that from working.